+++ verify/lib/mix/tasks/ua_inspector/verify.ex
@@ -25,6 +25,7 @@ defmodule Mix.Tasks.UaInspector.Verify do
   defp compare(%{client: _} = testcase, %{client: _} = result) do
     # regular user agent
     testcase.user_agent == result.user_agent &&
+      testcase.browser_family == result.browser_family &&
       testcase.client == maybe_from_struct(result.client) &&
       testcase.device == maybe_from_struct(result.device) &&
       testcase.os == maybe_from_struct(result.os)